将redis应用在部分的查询上减少数据库访问，顺带做了一些代码规范

+++ src/main/java/top/imzdx/storequeue/controller/GoodsController.java
@@ -56,7 +56,7 @@ public class GoodsController {
     }
 
     @GetMapping("{id}")//根据id获取商品
-    public Result getGoods(@PathVariable int id) {
+    public Result getGoods(@PathVariable long id) {
         Goods i = goodsService.getGoods(id);
         if (i != null) {
             return new ResultTools().success("获取成功", i);
